Add CoordinateElement with its CoordinateAvatar, CoordinateButtons and CoordinateReadings components
CoordinateReadings shows the coordinate's X (Longitude) and Y (Latitude).
CoordinateAvatar shows the coordinate's index in CoordinateList and allows the user to click on it to show/center the coordinate on the map view.
CoordinateButtons provides button to start/stop editing the coordinate, to delete the coordinate. The last button shows a menu in which the user can select to add a new coordinate above/below the current coordinate.