Add CoordinateList and its CoordinateListHeader component
CoordinateList shows a scrollable list of CoordinateElements. If the list is empty a text message appears in it's place explaining how to add the first coordinate.
CoordinateListHeader shows the title and a button that opens a menu with start/stop editing all coordinates, import/export, help and clear all coordinates functions.