Add IndexedMarker, DraggableMarker and MarkerMenuContext components
IndexedMarker renders a marker on the map with a on hover tooltip showing a markers index. On right mouse click it shows a context menu.
DraggableMarker is similar to IndexedMarker, but it can also be dragged on the map to change its position.
MarkerMenuContext is used to show a menu for markers. The menu allows to start/stop editing the coordinate, add a new coordinate before/after the current one, find and show the coordinate in the CoordinateList and delete the coordinate.