Add TrailMap, TrailRender, TrailMapEvents and MapMenuContext components
TrailMap component shows a "react-leaflet" map and uses TrailMapEvents for functionality and TrailRender for trail(coordinate list) rendering.
TrailRender renders the trail start and end markers, every coordinate's indicators, draggable markers when the coordinate is being edited and the trail path which is curved using the Bezier curve algorithm.
TrailMapEvents allows adding new coordinates on the map by holding 'ALT' button and pressing left click button on the map. It also allows to right click on the map and add a new coordinate there by clicking on the opened menu only option.
MapMenuContext allows opening that right click menu on the map.